refactor and modularize fragment, repository and viewmodel classes
delete unused classes

update main activity to use updated classes
