+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/BaseViewModel.java
@@ -2,28 +2,29 @@ package com.sdzshn3.android.news247.ViewModel;
 
 import android.app.Application;
 
-import com.sdzshn3.android.news247.Repositories.TopHeadlinesRepository;
+import com.sdzshn3.android.news247.Repositories.BaseRepository;
 import com.sdzshn3.android.news247.Retrofit.Article;
 
 import java.util.List;
 
+import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
-public class TopHeadlinesViewModel extends AndroidViewModel {
+public class BaseViewModel extends AndroidViewModel {
 
-    private TopHeadlinesRepository repository;
+    private BaseRepository repository;
 
-    public TopHeadlinesViewModel(Application application) {
+    public BaseViewModel(@NonNull Application application) {
         super(application);
-        repository = new TopHeadlinesRepository();
-    }
-
-    public void refresh() {
-        repository.refresh();
+        repository = new BaseRepository();
     }
 
     public LiveData<List<Article>> getData() {
         return repository.getData();
+    }
+
+    public void refresh() {
+        repository.refresh();
     }
 }
