--- app/src/main/java/com/sdzshn3/android/news247/ViewModel/EntertainmentViewModel.java
@@ -1,30 +1,0 @@
-package com.sdzshn3.android.news247.ViewModel;
-
-import android.app.Application;
-
-import com.sdzshn3.android.news247.Repositories.EntertainmentRepository;
-import com.sdzshn3.android.news247.Retrofit.Article;
-
-import java.util.List;
-
-import androidx.annotation.NonNull;
-import androidx.lifecycle.AndroidViewModel;
-import androidx.lifecycle.LiveData;
-
-public class EntertainmentViewModel extends AndroidViewModel {
-
-    private EntertainmentRepository repository;
-
-    public EntertainmentViewModel(@NonNull Application application) {
-        super(application);
-        repository = new EntertainmentRepository();
-    }
-
-    public LiveData<List<Article>> getData() {
-        return repository.getData();
-    }
-
-    public void refresh() {
-        repository.refresh();
-    }
-}
